+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/OverloadMethodsDeclarationOrderCheckTest.java
@@ -35,7 +35,9 @@ public class OverloadMethodsDeclarationOrderCheckTest
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "coding" + File.separator + filename);
+                + "coding" + File.separator
+                + "overloadmethoddeclaration" + File.separator
+                + filename);
     }
 
     @Test
